refactor builtin wrapper, etc
