+++ web/src/beta/features/Navbar/index.tsx
@@ -80,4 +80,5 @@ const Wrapper = styled.div`
   background: ${({ theme }) => theme.bg[0]};
   border-bottom: 0.5px solid ${({ theme }) => theme.outline.weak};
   box-shadow: 0px 0px 10px rgba(0, 0, 0, 0.5);
+  z-index: 1;
 `;
